File name change and added .bash_profile backup with timestamp
In this update the file names were updated to proper PEP 8 format. Along with this the file locations have been changed to $HOME to allow other user to utilize funtionality. Finally, the backup function has been added to the .bash_profile update (when old .bash_profile is written over a backup is stored on the ~ directory under .bash_profile_YYYY_MM_DD)